.claude/worktrees/agent-ab718158/.claude/worktrees/agent-ae3c8263/pyproject.toml
[project]
name = "sales-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.116.1",
    "jinja2>=3.1.6",
    "python-multipart>=0.0.20",
    "uvicorn>=0.35.0",
    "httpx>=0.27.0",
    "openai>=1.0.0",
    "whisper-live @ git+https://github.com/collabora/WhisperLive.git",
]
